Revert "Allow concurrent exports (#2181)"
This reverts commit b2c60e751b298b0fd10b63edbef25376575eb18c.
