+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/telemetry/BatchSpanProcessor.java
@@ -232,28 +232,6 @@ public final class BatchSpanProcessor {
     }
 
     private CompletableResultCode forceFlush() {
-      CompletableResultCode overallResult = new CompletableResultCode();
-      CompletableResultCode workerResult = forceFlushWorker();
-      workerResult.whenComplete(
-          () -> {
-            if (!workerResult.isSuccess()) {
-              overallResult.fail();
-              return;
-            }
-            CompletableResultCode exporterResult = spanExporter.flush();
-            exporterResult.whenComplete(
-                () -> {
-                  if (exporterResult.isSuccess()) {
-                    overallResult.succeed();
-                  } else {
-                    overallResult.fail();
-                  }
-                });
-          });
-      return overallResult;
-    }
-
-    private CompletableResultCode forceFlushWorker() {
       CompletableResultCode flushResult = new CompletableResultCode();
       // we set the atomic here to trigger the worker loop to do a flush of the entire queue.
       if (flushRequested.compareAndSet(null, flushResult)) {
